PT5/requirements2.txt
absl-py==2.1.0
attrs==23.2.0
beautifulsoup4==4.12.2
cachetools==5.3.2
centroid-tracker==0.0.9
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
colorzero==2.0
contourpy==1.1.1
cycler==0.12.1
filelock==3.13.1
flatbuffers==23.5.26
fonttools==4.44.3
gdown==4.7.1
google-auth==2.26.2
google-auth-oauthlib==1.0.0
gpiozero==2.0
grpcio==1.60.0
idna==3.4
importlib-metadata==7.0.1
importlib-resources==6.1.1
imutils==0.5.4
Jetson.GPIO==2.1.6
kiwisolver==1.4.5
Markdown==3.5.2
MarkupSafe==2.1.4
matplotlib==3.7.4
mediapipe==0.10.9
numpy==1.24.4
nvidia-pyindex==1.0.9
oauthlib==3.2.2
onnx==1.15.0
onnxscript==0.1.0.dev20231120
opencv-contrib-python==4.9.0.80
opencv-python==4.8.1.78
packaging==23.2
pandas==2.0.3
Pillow==10.1.0

protobuf==3.20.3
psutil==5.9.6
py-cpuinfo==9.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pyparsing==3.1.1
pyrealsense2==2.54.2.5684
pyserial==3.5
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
RPi.GPIO==0.7.1
rsa==4.9
scipy==1.10.1
seaborn==0.13.0
six==1.16.0
sounddevice==0.4.6
soupsieve==2.5
tensorboard==2.14.0
tensorboard-data-server==0.7.2
thop==0.1.1.post2209072238

tqdm==4.66.1
typing_extensions==4.8.0
tzdata==2023.3
ultralytics==8.1.23
urllib3==2.1.0
Werkzeug==3.0.1
zipp==3.17.0
